+++ common/net/src/msg/server.rs
@@ -139,7 +139,7 @@ pub enum ServerGeneral {
     CharacterCreated(character::CharacterId),
     CharacterEdited(character::CharacterId),
     CharacterSuccess,
-    SpectatorSuccess(Vec3<f32>),
+    SpectatorSuccess(Vec3<f64>),
     //Ingame related
     GroupUpdate(comp::group::ChangeNotification<Uid>),
     /// Indicate to the client that they are invited to join a group
@@ -173,7 +173,7 @@ pub enum ServerGeneral {
     /// if corrections are made.
     SetViewDistance(u32),
     Outcomes(Vec<Outcome>),
-    Knockback(Vec3<f32>),
+    Knockback(Vec3<f64>),
     // Ingame related AND terrain stream
     TerrainChunkUpdate {
         key: Vec2<i32>,
@@ -207,7 +207,7 @@ pub enum ServerGeneral {
     WeatherUpdate(WeatherGrid),
     /// Suggest the client to spectate a position. Called after client has
     /// requested teleport etc.
-    SpectatePosition(Vec3<f32>),
+    SpectatePosition(Vec3<f64>),
 }
 
 impl ServerGeneral {
