+++ common/src/comp/body/item_drop.rs
@@ -13,7 +13,7 @@ use crate::{
 };
 use rand::prelude::*;
 use serde::{Deserialize, Serialize};
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 use vek::Vec3;
 
 make_case_elim!(
@@ -121,10 +121,10 @@ impl Body {
 
     pub fn mass(&self) -> Mass { Mass(2.0) }
 
-    pub fn dimensions(&self) -> Vec3<f32> { Vec3::new(0.0, 0.1, 0.0) }
+    pub fn dimensions(&self) -> Vec3<f64> { Vec3::new(0.0, 0.1, 0.0) }
 
     pub fn orientation(&self, rng: &mut impl Rng) -> Ori {
-        let random = rng.gen_range(-1.0..1.0f32);
+        let random = rng.gen_range(-1.0..1.0f64);
         let default = Ori::default();
         match self {
             Body::Tool(_) => default
@@ -133,8 +133,8 @@ impl Body {
                 .pitched_towards(
                     Dir::from_unnormalized(Vec3::new(
                         random,
-                        rng.gen_range(-1.0..1.0f32),
-                        rng.gen_range(-1.0..1.0f32),
+                        rng.gen_range(-1.0..1.0f64),
+                        rng.gen_range(-1.0..1.0f64),
                     ))
                     .unwrap_or_default(),
                 ),
